cmd/libsnap-confine-private, cmd/s-c: move helpers to test utils
Move test_argc_argv to test utils.
