+++ cmd/snap-confine/snap-confine-args.c
@@ -21,6 +21,7 @@
 
 #include "../libsnap-confine-private/utils.h"
 #include "../libsnap-confine-private/string-utils.h"
+#include "../libsnap-confine-private/test-utils.h"
 
 struct sc_args {
 	// The security tag that the application is intended to run with
